Add unary minus operator to TraceableString
On a standard String, the unary minus operator returns a possibly
deduplicated version of the object. Since Ruby 2.7, the Psych YAML
parser calls this operator on all String-like objects to be used
as mapping indices. This made such TraceableString:s turn into
regular String:s. Overload the operator with one that never
deduplicates but instead returns the TraceableString itself.

This means we will use slightly more memory than strictly required,
but with the data volumes we anticipate this is not a concern.