Add pj2ansible tool
In order to be able to use Ansible with Pallet Jack, we need to
export the Pallet Jack data to Ansible. Writing an inventory
plugin would be possible, but just like for Salt we don't want to
have to run the entire Pallet Jack machinery for every invocation
of our configuration management tool. Hence, this tool that
exports the host list, global variables, and host-specific
variables to YAML files ready to be picked up by Ansible.