Allow pj2ansible to write inventory using IPv4 addresses
When bootstrapping, or when the Ansible control node is outside of
the Pallet Jack-controlled network and cannot resolve names, it
is very useful to have the option of using IP addresses for Ansible
connections instead of FQDN:s.

Add a new flag -i/--use-ipv4 to pj2ansible to select this mode.