Restrict to Knot's "simplified YAML"
Knot's configuration format is not YAML, as it may look at first
glance, but "simplified YAML". The parser does not like what Ruby's
default YAML formatter exports, but we have no way of influencing
it short of building our own YAML tree walker, which is too much
work. Instead, just do the bare minimum of string manipulation
required to make the default configuration pass Knot's parser,
and hope that it continues working in the future.