Basic Ansible roles and scripts for maintaining an environment
This is a very minimal set of Ansible roles, Ansible configuration,
and ancillary scripts for maintaining an environment using Pallet
Jack. It is enough to get to the point where network installation
of more systems is possible, but contains nothing fancy. It also
hard-codes a lot of things that should be configurable, such as
keyboard mappings and root passwords. And the only OS supported
is CentOS 8.1.1911. But, in the spirit of minimum viable product,
this works for me.